Update PlanLevel import and move it to a separate file
